integration: extract test_prepared_statement
Test `test_prepared_statement()` is extracted into
`statements/prepared.rs`.
