+++ scylla/tests/integration/statements/mod.rs
@@ -1,6 +1,7 @@
 mod batch;
 mod consistency;
 mod execution_profiles;
+mod prepared;
 mod silent_prepare_query;
 mod skip_metadata_optimization;
 mod statement;
